scripts/requirements.txt
python-dotenv
requests
deluge-client
plexapi
psutil
speedtest-cli
